lab_2/requirements.txt
wolframclient
wolframwebengine
python-dotenv
fastapi[uvicorn]
uvicorn
